Function finchers_r2d2::current_thread::pool_endpoint

pub fn pool_endpoint<M>(pool: Pool<M>) -> PoolEndpoint<M> where
    M: ManageConnection, 

Create an endpoint which acquires the connection from the specified connection pool.

This endpoint will block the current thread during retrieving the connection from pool.